more buffer fixes

+++ viridian/reef/src/lib/crypto/asymmetric.rs
@@ -62,7 +62,7 @@ impl Asymmetric {
     fn hide_public_key(&self, public_key: &ByteBuffer) -> DynResult<ByteBuffer> {
         let mut state = Blake2bVar::new(SYMMETRIC_HASH_SIZE)?;
         let result = ByteBuffer::empty(N_SIZE + SYMMETRIC_HASH_SIZE);
-        let (n_number, hash) = result.split_buf(N_SIZE as isize);
+        let (n_number, hash) = result.split_buf(N_SIZE);
         get_rng().fill_bytes(&mut n_number.slice_mut());
         state.update(&n_number.slice());
         state.update(&self.seed_key);
